+++ core/src/main/java/com/arnyminerz/escalaralcoiaicomtat/core/ui/element/Chip.kt
@@ -1,4 +1,4 @@
-package com.arnyminerz.escalaralcoiaicomtat.ui.elements
+package com.arnyminerz.escalaralcoiaicomtat.core.ui.element
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
